handlebars: Move popover handlebars to a new popovers folder.
Organize related handlebars in a new folder together.
