+++ tools/linter_lib/custom_check.py
@@ -694,7 +694,7 @@ html_rules: List["Rule"] = [
             "templates/corporate/billing.html",
             "templates/corporate/upgrade.html",
             # Miscellaneous violations to be cleaned up
-            "web/templates/user_card_popover_avatar.hbs",
+            "web/templates/popovers/user_card/user_card_popover_avatar.hbs",
             "web/templates/confirm_dialog/confirm_subscription_invites_warning.hbs",
             "templates/zerver/reset_confirm.html",
             "templates/zerver/config_error.html",
