+++ web/src/add_stream_options_popover.js
@@ -1,6 +1,6 @@
 import $ from "jquery";
 
-import render_left_sidebar_stream_setting_popover from "../templates/left_sidebar_stream_setting_popover.hbs";
+import render_left_sidebar_stream_setting_popover from "../templates/popovers/left_sidebar_stream_setting_popover.hbs";
 
 import * as popover_menus from "./popover_menus";
 import * as settings_data from "./settings_data";
