+++ web/templates/compose_control_buttons.hbs
@@ -24,7 +24,7 @@
     </div>
     <div class="divider hide-sm show-md hide-mc">|</div>
     <div class="{{#if message_id}}hide-lg{{else}}hide-sm show-md hide-mc{{/if}}">
-        {{> compose_control_buttons_in_popover}}
+        {{> popovers/compose_control_buttons/compose_control_buttons_in_popover}}
     </div>
     <a role="button" class="compose_control_button compose_draft_button hide-sm" tabindex=0 href="#drafts" data-tooltip-template-id="compose_draft_tooltip_template">
         {{t 'Drafts' }}
