+++ web/templates/popovers/emoji/emoji_popover.hbs
