+++ web/templates/popovers/emoji/emoji_popover_emoji.hbs
