+++ web/templates/popovers/emoji/emoji_popover_search_results.hbs
