+++ web/templates/popovers/emoji/emoji_showcase.hbs
