+++ web/templates/popovers/user_card/user_card_popover_avatar.hbs
