+++ web/templates/popovers/user_group_info_popover.hbs
